+++ globals.h
@@ -1,12 +1,12 @@
-extern const double PI;
+extern const long double PI;
 
 extern const int nx;
 
 extern const int opt;
 
 extern const int nDesVar;
-extern const double a_geom, b_geom;
-extern const double h_geom, t1_geom, t2_geom;
+extern const long double a_geom, b_geom;
+extern const long double h_geom, t1_geom, t2_geom;
 
 extern const int fitnessFun;
 
@@ -16,20 +16,20 @@ extern const int gradientType;
 
 extern const int createTarget;
 
-extern const double gam;
-extern const double R;
-extern const double Cv;
+extern const long double gam;
+extern const long double R;
+extern const long double Cv;
 
-extern const double Min;
-extern const double Ttin; 
-extern const double ptin;
+extern const long double Min;
+extern const long double Ttin; 
+extern const long double ptin;
 
-extern const double pexit;
+extern const long double pexit;
 
-extern const double a2;
+extern const long double a2;
 
-extern const double CFL;
-extern const double conv;
+extern const long double CFL;
+extern const long double conv;
 
 extern const int maxIt;
 
@@ -41,4 +41,4 @@ extern const int StepScheme;
 
 extern const int FluxScheme;
 
-extern double Scalareps;
+extern long double Scalareps;
